Move applications to dedicated module
