+++ suricatalog/canned.py
@@ -1,21 +1,15 @@
 from __future__ import annotations
 
-import textwrap
 from pathlib import Path
-from typing import List, Callable, Any
+from typing import List
 import locale
 
-from rich.text import Text
-from textual import on
-from textual.app import App, ComposeResult
-from textual.widgets import Header, Footer, Digits, Pretty, DataTable
+from textual.app import App
 
+from suricatalog.apps import FlowApp, HostDataUse, TopUserApp, OneShotApp
 from suricatalog.filter import BaseFilter
-from suricatalog.log import get_events_from_eve
-from suricatalog.report import AggregatedFlowProtoReport, HostDataUseReport, TopUserAgents
 
 locale.setlocale(locale.LC_ALL, '')
-BASEDIR = Path(__file__).parent
 
 
 def get_one_shot_flow_table(
@@ -23,59 +17,11 @@ def get_one_shot_flow_table(
         eve: List[Path],
         data_filter: BaseFilter
 ) -> App:
+    flow_app = FlowApp(
+        eve=eve,
+        data_filter=data_filter
+    )
     logs = ' '.join(map(lambda x: str(x), eve))
-
-    class FlowApp(App):
-        BINDINGS = [
-            ("q", "quit_app", "Quit")
-        ]
-        ENABLE_COMMAND_PALETTE = False
-
-        def action_quit_app(self) -> None:
-            self.exit("Exiting Net-Flow now...")
-
-        def compose(self) -> ComposeResult:
-            yield Header()
-            alerts_tbl = DataTable()
-            alerts_tbl.show_header = True
-            alerts_tbl.add_column("Destination IP")
-            alerts_tbl.add_column("Port")
-            alerts_tbl.add_column("Count")
-            alerts_tbl.zebra_stripes = True
-            alerts_tbl.loading = True
-            alerts_tbl.cursor_type = 'row'
-            alerts_tbl.tooltip = textwrap.dedent("""
-            Network flow details
-            """)
-            yield alerts_tbl
-            yield Footer()
-
-        def on_mount(self) -> None:
-            alerts_tbl = self.query_one(DataTable)
-            alert_cnt = 0
-            afr = AggregatedFlowProtoReport()
-            for event in get_events_from_eve(
-                    eve_files=eve,
-                    data_filter=data_filter):
-                if not data_filter.accept(event):
-                    continue
-                afr.ingest_data(event)
-            for (dest_ip_port, cnt) in afr.port_proto_count.items():
-                alerts_tbl.add_row(
-                    dest_ip_port[0],
-                    str(dest_ip_port[1]),
-                    Text(str(cnt), style="italic #03AC13", justify="right")
-                )
-                alert_cnt += 1
-            alerts_tbl.loading = False
-
-        @on(DataTable.HeaderSelected)
-        def on_header_clicked(self, event: DataTable.HeaderSelected):
-            alerts_tbl: DataTable = event.data_table
-            if event.column_index < 2:
-                alerts_tbl.sort(event.column_key)
-
-    flow_app = FlowApp()
     flow_app.title = f"SuricataLog FLOW protocol, logs={logs}"
     return flow_app
 
@@ -85,38 +31,11 @@ def get_host_data_use(
         data_filter: BaseFilter,
         ip_address: any
 ) -> App:
-    host_data_user_report = HostDataUseReport()
-
-    class HostDataUse(App):
-        BINDINGS = [
-            ("q", "quit_app", "Quit")
-        ]
-        CSS_PATH = BASEDIR.joinpath('css').joinpath('canned.tcss')
-        ENABLE_COMMAND_PALETTE = False
-
-        def action_quit_app(self) -> None:
-            self.exit("Exiting Net-Flow now...")
-
-        def compose(self) -> ComposeResult:
-            yield Header()
-            digits = Digits(id="netflow")
-            digits.loading = True
-            digits.tooltip = textwrap.dedent("""
-            Net FLow in bytes.
-            """)
-            yield digits
-            yield Footer()
-
-        def on_mount(self) -> None:
-            for event in get_events_from_eve(
-                    eve_files=eve_files,
-                    data_filter=data_filter):
-                host_data_user_report.ingest_data(event, ip_address)
-            digits = self.query_one('#netflow', Digits)
-            digits.update(f"{host_data_user_report.bytes:n} bytes")
-            digits.loading = False
-
-    hdu = HostDataUse()
+    hdu = HostDataUse(
+        eve=eve_files,
+        data_filter=data_filter,
+        ip_address=ip_address
+    )
     hdu.title = f"SuricataLog Net-Flow for: {ip_address}"
     return hdu
 
@@ -125,35 +44,10 @@ def get_agents(
         eve_files: List[Path],
         data_filter: BaseFilter
 ) -> App:
-    top_user_agents = TopUserAgents()
-
-    class TopUserApp(App):
-        BINDINGS = [
-            ("q", "quit_app", "Quit")
-        ]
-        CSS_PATH = BASEDIR.joinpath('css').joinpath('canned.tcss')
-        ENABLE_COMMAND_PALETTE = False
-
-        def action_quit_app(self) -> None:
-            self.exit("Exiting Top user now...")
-
-        def compose(self) -> ComposeResult:
-            yield Header()
-            pretty = Pretty("", id="agent")
-            pretty.loading = True
-            yield pretty
-            yield Footer()
-
-        def on_mount(self) -> None:
-            pretty = self.query_one("#agent", Pretty)
-            for event in get_events_from_eve(
-                    eve_files=eve_files,
-                    data_filter=data_filter):
-                top_user_agents.ingest_data(event)
-            pretty.update(top_user_agents.agents)
-            pretty.loading = False
-
-    top_user_app = TopUserApp()
+    top_user_app = TopUserApp(
+        eve=eve_files,
+        data_filter=data_filter
+    )
     top_user_app.title = f"SuricataLog User Agents"
     return top_user_app
 
@@ -161,32 +55,12 @@ def get_agents(
 def get_capture(
         *,
         eve: List[Path],
-        retriever: Callable = get_events_from_eve,
         data_filter: BaseFilter,
         title: str
 ) -> App:
-    class OneShotApp(App):
-        BINDINGS = [
-            ("q", "quit_app", "Quit")
-        ]
-        ENABLE_COMMAND_PALETTE = False
-
-        def action_quit_app(self) -> None:
-            self.exit(f"Exiting {title} now...")
-
-        def compose(self) -> ComposeResult:
-            yield Header()
-            pretty = Pretty("", id='events')
-            pretty.loading = True
-            yield pretty
-            yield Footer()
-
-        def on_mount(self):
-            pretty = self.query_one('#events', Pretty)
-            events: List[Any] = [single_alert for single_alert in retriever(eve_files=eve, data_filter=data_filter)]
-            pretty.update(events)
-            pretty.loading = False
-
-    one_shot_app = OneShotApp()
+    one_shot_app = OneShotApp(
+        eve=eve,
+        data_filter=data_filter
+    )
     one_shot_app.title = f"{title}"
     return one_shot_app
